Implement the Zoetrope Arduino project in Clojure.
Using `clj-firmata` I used a declarative approach to solving the problem.

By having the pin information in a simple map, I was able to use that
information in the setup function which then requires reasoning on input vs
output. The event-based approach worked around this neatly and the input logic
could be wrapped in `hook-up-on-event`, which sets up async channels, listening
to incoming events from the board publisher.

This implementation has a shortcoming in that changing the value from the
potentiometer does not change the speed in real time, and only when the
stop-start button is toggled.